19102017_A
Changed the scoring system! So now if the ball exits the court from anywhere (left, right, bottom), the player loses one point, and if score reaches -10 game is over.
Ugh just realized it's not a good idea for each collision to score +1 point and add to the same int, then the score will be a sum of each losing round and it'll bring the score from negative to positive, so I have:
a) still have to scoring systems, perhaps left for gaining (green?), right for losing (red), if either reaches 10, game over
b) the collision scoring initializes with a 0 each time serving starts again, so the way we win is just to successfully 'fetch' the ball 10 times before we lose it over 10 times!